==> code/05_explorers_questions.R <==
#05_Explorer’s questions

#1-In retrospect, is DOMA a representative station? Why do you think its behaviour is so different than the other stations? 
#Doma's very high altitude, and has low levels of runoff. I believe the data is still relevant, and the station despite having significantly less runoff than others, should still be factored. Once there is consecutively null data from this station, then it should be deleted.Despite the effects of interacting with the water flow, humans should still count this location. How to handle this situation when newer scientists arriving in the field and seeing the data sets, without the station, if they are unaware of the missing catchment? 

#2-In our analysis, we have used only river runoff. Precipitation is a factor strongly linked with runoff. Can you perform a similar analysis (boxplots and regression) for precipitation? Precipitation data averaged over the whole Rhine region can be found in the file precip_day.rds in folder data. What do you observe?

  
#3-What are your thoughts about the changes in Rhine runoff after completing EDA?
#definitely seeing the changes over time, and see how runoff can depict changes in global condition. This being shown in the figures we got from actual data was exhilerating.

#4-Which are some future analyses or other factors that should be examined? I believe the temperature being recorded, and the annual mean increase/decrease would be a great table.
